Ajuste funcional simulacao

+++ trabalho-sekeff-main/src/main/java/com/example/simulation/traffic/Roteirizador.java
@@ -2,7 +2,6 @@ package com.example.simulation.traffic;
 
 import java.io.Serializable;
 
-import com.example.simulation.datastructure.Fila;
 import com.example.simulation.datastructure.LinkedList;
 import com.example.simulation.graph.Grafo;
 import com.example.simulation.graph.Intersecao;
@@ -16,15 +15,15 @@ public class Roteirizador implements Serializable {
         this.grafo = grafo;
     }
 
-    public LinkedList<Integer> calcularRota(int origemId, int destinoId) {
+    public LinkedList<Long> calcularRota(int origemId, int destinoId) {
         Intersecao origem = grafo.obterIntersecaoPorId(origemId);
         Intersecao destino = grafo.obterIntersecaoPorId(destinoId);
 
-        LinkedList<Integer> caminho = Dijkstra.encontrarMenorCaminho(grafo, origem, destino);
-        LinkedList<Integer> rota = new LinkedList<>();
+        LinkedList<Long> caminho = Dijkstra.encontrarMenorCaminho(grafo, origem, destino);
+        LinkedList<Long> rota = new LinkedList<>();
 
         while (!caminho.isEmpty()) {
-            Integer elem = caminho.getPrimeiro();
+            Long elem = caminho.getPrimeiro();
             caminho.removeFirst();
             rota.add(elem);
         }
